backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10
email-validator==2.1.0
psutil==5.9.6
prometheus-client==0.19.0
pydantic==2.5.0
python-jose[cryptography]>=3.3.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
celery==5.3.4
redis==5.0.1
lxml==4.9.3
pandas==2.1.3
boto3==1.34.0
paramiko==3.4.0
pytest==7.4.3
pytest-asyncio==0.21.1
kombu==5.3.4
botocore==1.34.0
defusedxml==0.7.1
werkzeug==3.0.1
slowapi==0.1.9
requests==2.31.0
pytz==2023.3